+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.7] - 2018-09-20
+
 ### Fixed
 
 - Fix refreshing buffer count in the minibuffer right label when a buffer is